more stuff

+++ public_html/bundles/site/js/home/home.js
@@ -3,11 +3,27 @@ $(function () {
     $('.questionnaire-delete').click(function () {
         var id = $(this).data('id');
 
-        $.ajax({
-            type: "DELETE",
-            url: "/moderate/questionnaire/" + id,
-            success: function () {
-                location.reload();
+        bootbox.dialog({
+            message: "Ar jūs tikrai norite pašalinti šį klausimyną?",
+            title: "Pašalinti",
+            buttons: {
+                yes: {
+                    label: "Pašalinti",
+                    className: "btn-danger",
+                    callback: function () {
+                        $.ajax({
+                            type: "DELETE",
+                            url: "/moderate/questionnaire/" + id,
+                            success: function () {
+                                location.reload();
+                            }
+                        });
+                    }
+                },
+                no: {
+                    label: "Atšaukti",
+                    className: "btn-success"
+                }
             }
         });
     });
